bugs(code fixed)
