+++ frontend/src/App.jsx
@@ -1,40 +1,23 @@
-// layout component that wraps pages with sidebar
-import React, { useState } from "react";
-import { Box } from "@mui/material";
-import Sidebar from "./components/Sidebar";
+import React from "react";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Layout from "./components/Layout";
 
-const Layout = ({ children }) => {
-  const [collapsed, setCollapsed] = useState(false);
+// actual pages
+import Home from "./pages/Home";
+import TimeSeries from "./pages/TimeSeries";
+import NotFound from "./pages/NotFound";
 
+const App = () => {
   return (
-    <Box sx={styles.appContainer}>
-      {/* pass collapsed + toggle down to Sidebar */}
-      <Sidebar collapsed={collapsed} setCollapsed={setCollapsed} />
-
-      <Box
-        component="main"
-        sx={{
-          ...styles.mainContent,
-          ml: collapsed ? "80px" : "280px", // shift depending on sidebar
-          transition: "margin-left 0.3s ease",
-        }}
-      >
-        {children}
-      </Box>
-    </Box>
+    <BrowserRouter>
+      <Layout>
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/timeseries" element={<TimeSeries />} />
+        </Routes>
+      </Layout>
+    </BrowserRouter>
   );
 };
 
-export default Layout;
-
-const styles = {
-  appContainer: {
-    display: "flex",
-    minHeight: "100vh",
-    bgcolor: "#f5f5f5",
-  },
-  mainContent: {
-    flexGrow: 1,
-    p: 4,
-  },
-};
+export default App;
